added delay before display init, removed unused file
On one of the ThinkTinys the display would only work after turning it
on-off-on. Adding a small delay before makes it work consistently,
probably because the voltage has time to stabilise.

Removed unused/empty autogenerated slpcontrol file to get the
required space. There's some more potential to save space there.